requirements.txt
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
comm==0.1.3
cryptography==41.0.2
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.2
executing==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
ipykernel==6.24.0
ipython==8.14.0
jedi==0.18.2
jupyter_client==8.3.0
jupyter_core==5.3.1
kaitaistruct==0.10
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.24.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==3.9.1
prompt-toolkit==3.0.39
psutil==5.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pycparser==2.21
pydivert==2.1.0
Pygments==2.15.1
pyOpenSSL==23.2.0
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
selenium==4.10.0
selenium-wire==5.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.1
trio-websocket==0.10.3
tzdata==2023.3
Unidecode==1.3.6
urllib3==2.0.3
wcwidth==0.2.6
webdriver-manager==3.8.6
wsproto==1.2.0
zstandard==0.21.0
